+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# REPLACE_GLOBALLY_WITH_NEXT_VERSION
+# 3.1.0
 - PPI-246 - Added option for advanced logging
 - PPI-251 - Fix authorized transaction state on delayed payment collection
 - PPI-276 - Fix multiple unnecessary requests in the background with PayPal Plus